Refactor Lua classes to individual files
