+++ src/lua_env/lua_class/lua_class_item.hpp
@@ -1,0 +1,15 @@
+#pragma once
+#include "../lua_api/lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace LuaItem
+{
+void bind(sol::state&);
+}
+
+}
+}
