+++ src/lua_env/lua_class/lua_class_position.hpp
@@ -1,0 +1,15 @@
+#pragma once
+#include "../lua_api/lua_api_common.hpp"
+
+namespace elona
+{
+namespace lua
+{
+
+namespace LuaPosition
+{
+void bind(sol::state&);
+}
+
+}
+}
